requirements.txt
# ─── Web framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Data / ML ───────────────────────────────────────────────────────────────
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.26.4
joblib==1.3.2
scipy==1.11.4

# ─── Deep learning ───────────────────────────────────────────────────────────
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0

# ─── LLM integration ─────────────────────────────────────────────────────────
openai==1.30.1
httpx==0.27.0

# ─── Validation ──────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Environment ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Async / File handling ───────────────────────────────────────────────────
aiofiles==23.2.1

# ─── Rate limiting ───────────────────────────────────────────────────────────
slowapi==0.1.9

# ─── Logging ─────────────────────────────────────────────────────────────────
loguru==0.7.2

# ─── Language detection ──────────────────────────────────────────────────────
langdetect==1.0.9

# ─── HTTP requests ───────────────────────────────────────────────────────────
requests==2.31.0
